requirements.txt
django ~= 3.0
djangorestframework
djangorestframework-jwt
django-rest-swagger
